created GhostPlayer object to ease merging later
