+++ src/main/java/nl/tudelft/jpacman/level/MultiGhostPlayerCollisions.java
@@ -3,10 +3,6 @@ package nl.tudelft.jpacman.level;
 import nl.tudelft.jpacman.board.Unit;
 import nl.tudelft.jpacman.npc.ghost.Ghost;
 
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-
 /**
  * Created by helldog136 on 26/02/16.
  */
@@ -15,17 +11,17 @@ public class MultiGhostPlayerCollisions extends PlayerCollisions{
     @Override
     public void collide(Unit mover, Unit collidedOn) {
 
-        if (mover instanceof GhostPlayer) {
-            playerColliding((GhostPlayer) mover, collidedOn);
+        if (mover instanceof HunterGhostPlayer) {
+            playerColliding((HunterGhostPlayer) mover, collidedOn);
         }
         else if (mover instanceof Ghost) {
             ghostColliding((Ghost) mover, collidedOn);
         }
     }
     
-    protected void playerColliding(GhostPlayer player, Unit collidedOn) {
-        if (collidedOn instanceof GhostPlayer) {
-            playerVersusPlayer(player, (GhostPlayer) collidedOn);
+    protected void playerColliding(HunterGhostPlayer player, Unit collidedOn) {
+        if (collidedOn instanceof HunterGhostPlayer) {
+            playerVersusPlayer(player, (HunterGhostPlayer) collidedOn);
         }
 
         if (collidedOn instanceof Pellet) {
@@ -33,7 +29,7 @@ public class MultiGhostPlayerCollisions extends PlayerCollisions{
         }
     }
 
-    private void playerVersusPlayer(GhostPlayer player1, GhostPlayer player2) {
+    private void playerVersusPlayer(HunterGhostPlayer player1, HunterGhostPlayer player2) {
         //TODO blocking
         if(player1.isHunter() || player2.isHunter()) {
             if(player1.isHunter()) {
@@ -52,7 +48,7 @@ public class MultiGhostPlayerCollisions extends PlayerCollisions{
      * @param player The player involved in the collision.
      * @param pellet The pellet involved in the collision.
      */
-    public void playerVersusPellet(GhostPlayer player, Pellet pellet) {
+    public void playerVersusPellet(HunterGhostPlayer player, Pellet pellet) {
         if(player.isActive()) {
             pellet.leaveSquare();
             player.addPoints(pellet.getValue());
